add comment

+++ lib/data_process.py
@@ -175,17 +175,20 @@ class ReconstructionDataProcess(DataProcess):
 
 def kill_processes(queue, processes):
     print('Signal processes')
-    for p in processes:
-        p.shutdown()
+    if processes:
+        for p in processes:
+            p.shutdown()
 
     print('Empty queue')
-    while not queue.empty():
-        time.sleep(0.5)
-        queue.get(False)
+    if queue:
+        while not queue.empty():
+            time.sleep(0.5)
+            queue.get(False)
 
     print('kill processes')
-    for p in processes:
-        p.terminate()
+    if processes:
+        for p in processes:
+            p.terminate()
 
 
 def make_data_processes(queue, data_paths, num_workers, repeat=True, train=True):
